requirements.txt
asgiref==3.4.1
astroid==2.4.2
autopep8==1.6.0
Babel==2.9.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.0
defusedxml==0.7.1
dj-rest-auth==2.1.11
Django==3.2.10
django-allauth==0.44.0
django-cors-headers==3.10.1
django-crum==0.7.9
django-model-utils==4.2.0
django-money==2.1
djangorestframework==3.13.0
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
isort==5.6.4
itypes==1.2.0
Jinja2==3.0.3
lazy-object-proxy==1.4.3
MarkupSafe==2.0.1
mccabe==0.6.1
oauthlib==3.1.1
packaging==21.3
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
pylint==2.6.0
pyparsing==3.0.6
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
six==1.15.0
sqlparse==0.4.2
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.7
wrapt==1.12.1
